Added seleniumTests to codeCov?
