+++ .github/workflows/tests.yml
@@ -32,6 +32,9 @@ jobs:
       - run: python3 -m pip install -r requirements.txt
       - run: pip install -e ".[dev]"
       - run: python3 -m pytest --cov=core --cov=website --cov=Slurm --cov-report=xml tests/test_PeakLearner.py
+      - run: python3 -m pytest --cov=core --cov=website --cov=Slurm --cov-report=xml --cov-append tests/core/test_jobTests.py
+      - uses: nanasess/setup-chromedriver@master
+      - run: python3 -m pytest --cov=core --cov=website --cov=Slurm --cov-report=xml --cov-append seleniumTests/test_withSelenium.py
       - name: Upload coverage to Codecov
         uses: codecov/codecov-action@v1
         with:
@@ -42,40 +45,10 @@ jobs:
           fail_ci_if_error: true
           path_to_write_report: ./codecov_report.txt
           verbose: true
-  seleniumTests:
-    name: Selenium Tests Runner
-    runs-on: ubuntu-latest
-    steps:
-      - uses: actions/checkout@v2
-      - name: Set up Python 3.8
-        uses: actions/setup-python@v2
-        with:
-          python-version: 3.8
-      - uses: nanasess/setup-chromedriver@master
-      - run: python3 -m pip install -U pip
-      - run: sudo apt-get update
-      - run: sudo apt-get -y install libdb-dev libdb++-dev samtools r-base
-      - run: sudo Rscript -e 'install.packages("data.table")'
-      - run: mkdir bin/
-      - run: wget http://hgdownload.soe.ucsc.edu/admin/exe/linux.x86_64.v385/bigWigSummary
-        working-directory: bin/
-      - run: wget http://hgdownload.soe.ucsc.edu/admin/exe/linux.x86_64.v385/bigWigToBedGraph
-        working-directory: bin/
-      - run: sudo chmod a+x bigWigSummary
-        working-directory: bin/
-      - run: sudo chmod a+x bigWigToBedGraph
-        working-directory: bin/
-      - run: python --version
-      - run: ./setup.sh
-        working-directory: jbrowse/jbrowse/
-      - run: python3 -m pip install numpy==1.20.3
-      - run: python3 -m pip install -r requirements.txt
-      - run: pip install -e ".[dev]"
-      - run: python3 -m pytest tests/test_withSelenium.py
   deployOpenAPI:
     name: Generates the OpenAPI docs website
     runs-on: ubuntu-latest
-    needs: [unitTests, seleniumTests]
+    needs: unitTests
     if: github.ref == 'refs/heads/main'
     steps:
       - uses: actions/checkout@v2
