+++ exampleCfg.cfg
@@ -10,4 +10,5 @@ doidlepredictions = False
 timebetween = 600
 numchanges = 10
 minlabeledregions = 20
+maxJobsToSpawn = 20
